==> 03a_regression_no_resample.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler  
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score,accuracy_score, precision_score, recall_score, f1_score
from sklearn.model_selection import KFold

df = pd.read_csv('cleaned_movies.csv')

X = df.drop(columns=['genres', 'original_language', 'popularity_class', 'popularity']).copy()
y = df['popularity']


print(X.dtypes)

df_all_results = pd.DataFrame()

# data for discretization of results

bins = [0, 10, 20, 30, 40, 50, 60, 70, 80, 90, y.max()]
labels = list(range(10))

# SVR

from sklearn.svm import SVR

pipeline = Pipeline([
    ('scaler', StandardScaler()),
    ('svr', SVR())
])

kf = KFold(n_splits=10, shuffle=True, random_state=42)

mae_scores = []
rmse_scores = []
r2_scores = []

acc_scores = []
prec_scores = []
rec_scores = []
f1_scores = []

for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]

    pipeline.fit(X_train, y_train)
    y_pred = pipeline.predict(X_test)

    # MAE
    mae = mean_absolute_error(y_test, y_pred)
    mae_scores.append(mae)

    # RMSE
    rmse = np.sqrt(mean_squared_error(y_test, y_pred))
    rmse_scores.append(rmse)
    
    # R2
    r2 = r2_score(y_test, y_pred)
    r2_scores.append(r2)
    
    
    y_pred_disc = pd.cut(y_pred, bins=bins, labels=labels, include_lowest=True)
    y_test_disc = pd.cut(y_test, bins=bins, labels=labels, include_lowest=True)
    
    mask = (~pd.isna(y_pred_disc)) & (~pd.isna(y_test_disc))
    y_pred_disc = y_pred_disc[mask]
    y_test_disc = y_test_disc[mask]

    acc = accuracy_score(y_test_disc, y_pred_disc)
    prec = precision_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    rec = recall_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    f1 = f1_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    
    acc_scores.append(acc)
    prec_scores.append(prec)
    rec_scores.append(rec)
    f1_scores.append(f1)


print("\nMean Absolute Error (per fold):", mae_scores)
print("Mean MAE:", np.mean(mae_scores))

print("\nRoot Mean Squared Error (per fold):", rmse_scores)
print("Mean RMSE:", np.mean(rmse_scores))

print("\nR2 - Coefficient of Determination (per fold):", r2_scores)
print("Mean R2:", np.mean(r2_scores))

print("\n--- Summary after discretization---")
print(f"Mean Accuracy: {np.mean(acc_scores):.4f}")
print(f"Mean Precision: {np.mean(prec_scores):.4f}")
print(f"Mean Recall: {np.mean(rec_scores):.4f}")
print(f"Mean F1-score: {np.mean(f1_scores):.4f}")


# saving results to df
df_results = pd.DataFrame({
    'MAE': mae_scores,
    'RMSE': rmse_scores,
    'R2': r2_scores
})
df_results.to_csv("3a_regression_no_resample_SVR.csv", index=False)

# saving results to df (for comparison with classification)
df_results = pd.DataFrame({
    'acc': acc_scores,
    'prec': prec_scores,
    'rec': rec_scores,
    'f1': f1_scores
})
df_results.to_csv("3a_classification_no_resample_SVR.csv", index=False)

# KNN

from sklearn.neighbors import KNeighborsRegressor

pipeline = Pipeline([
    ('scaler', StandardScaler()),
    ('knn', KNeighborsRegressor(n_neighbors=5))
])

kf = KFold(n_splits=10, shuffle=True, random_state=42)

mae_scores = []
rmse_scores = []
r2_scores = []

acc_scores = []
prec_scores = []
rec_scores = []
f1_scores = []

for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]

    pipeline.fit(X_train, y_train)
    y_pred = pipeline.predict(X_test)

    mae = mean_absolute_error(y_test, y_pred)
    mae_scores.append(mae)

    rmse = np.sqrt(mean_squared_error(y_test, y_pred))
    rmse_scores.append(rmse)
    
    r2 = r2_score(y_test, y_pred)
    r2_scores.append(r2)
    
    y_pred_disc = pd.cut(y_pred, bins=bins, labels=labels, include_lowest=True)
    y_test_disc = pd.cut(y_test, bins=bins, labels=labels, include_lowest=True)
    
    mask = (~pd.isna(y_pred_disc)) & (~pd.isna(y_test_disc))
    y_pred_disc = y_pred_disc[mask]
    y_test_disc = y_test_disc[mask]

    acc = accuracy_score(y_test_disc, y_pred_disc)
    prec = precision_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    rec = recall_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    f1 = f1_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    
    acc_scores.append(acc)
    prec_scores.append(prec)
    rec_scores.append(rec)
    f1_scores.append(f1)



print("\nMean Absolute Error (per fold):", mae_scores)
print("Mean MAE:", np.mean(mae_scores))

print("\nRoot Mean Squared Error (per fold):", rmse_scores)
print("Mean RMSE:", np.mean(rmse_scores))

print("\nR2 - Coefficient of Determination (per fold):", r2_scores)
print("Mean R2:", np.mean(r2_scores))

print("\n--- Summary after discretization---")
print(f"Mean Accuracy: {np.mean(acc_scores):.4f}")
print(f"Mean Precision: {np.mean(prec_scores):.4f}")
print(f"Mean Recall: {np.mean(rec_scores):.4f}")
print(f"Mean F1-score: {np.mean(f1_scores):.4f}")

# saving resoults to df
df_results = pd.DataFrame({
    'MAE': mae_scores,
    'RMSE': rmse_scores,
    'R2': r2_scores
})
df_results.to_csv("3a_regression_no_resample_KNN.csv", index=False)

# saving results to df (for comparison with classification)
df_results = pd.DataFrame({
    'acc': acc_scores,
    'prec': prec_scores,
    'rec': rec_scores,
    'f1': f1_scores
})
df_results.to_csv("3a_classification_no_resample_KNN.csv", index=False)

# DECISION TREE

from sklearn.tree import DecisionTreeRegressor

pipeline = Pipeline([
    ('tree', DecisionTreeRegressor(random_state=42))
])

kf = KFold(n_splits=10, shuffle=True, random_state=42)

mae_scores = []
rmse_scores = []
r2_scores = []

acc_scores = []
prec_scores = []
rec_scores = []
f1_scores = []

for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]

    pipeline.fit(X_train, y_train)
    y_pred = pipeline.predict(X_test)

    mae = mean_absolute_error(y_test, y_pred)
    mae_scores.append(mae)

    rmse = np.sqrt(mean_squared_error(y_test, y_pred))
    rmse_scores.append(rmse)

    r2 = r2_score(y_test, y_pred)
    r2_scores.append(r2)
    
    y_pred_disc = pd.cut(y_pred, bins=bins, labels=labels, include_lowest=True)
    y_test_disc = pd.cut(y_test, bins=bins, labels=labels, include_lowest=True)
    
    mask = (~pd.isna(y_pred_disc)) & (~pd.isna(y_test_disc))
    y_pred_disc = y_pred_disc[mask]
    y_test_disc = y_test_disc[mask]

    acc = accuracy_score(y_test_disc, y_pred_disc)
    prec = precision_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    rec = recall_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    f1 = f1_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    
    acc_scores.append(acc)
    prec_scores.append(prec)
    rec_scores.append(rec)
    f1_scores.append(f1)


print("\nMean Absolute Error (per fold):", mae_scores)
print("Mean MAE:", np.mean(mae_scores))

print("\nRoot Mean Squared Error (per fold):", rmse_scores)
print("Mean RMSE:", np.mean(rmse_scores))

print("\nR2 - Coefficient of Determination (per fold):", r2_scores)
print("Mean R2:", np.mean(r2_scores))

print("\n--- Summary after discretization---")
print(f"Mean Accuracy: {np.mean(acc_scores):.4f}")
print(f"Mean Precision: {np.mean(prec_scores):.4f}")
print(f"Mean Recall: {np.mean(rec_scores):.4f}")
print(f"Mean F1-score: {np.mean(f1_scores):.4f}")

# saving resoults to df
df_results = pd.DataFrame({
    'MAE': mae_scores,
    'RMSE': rmse_scores,
    'R2': r2_scores
})
df_results.to_csv("3a_regression_no_resample_DT.csv", index=False)

# saving results to df (for comparison with classification)
df_results = pd.DataFrame({
    'acc': acc_scores,
    'prec': prec_scores,
    'rec': rec_scores,
    'f1': f1_scores
})
df_results.to_csv("3a_classification_no_resample_DT.csv", index=False)

# MULTILAYER PERCEPTRON (MLP)

from sklearn.neural_network import MLPRegressor

pipeline = Pipeline([
    ('scaler', StandardScaler()),
    ('mlp', MLPRegressor(hidden_layer_sizes=(100, 50), max_iter=1000, early_stopping=True, random_state=42))
])

kf = KFold(n_splits=10, shuffle=True, random_state=42)

mae_scores = []
rmse_scores = []
r2_scores = []

acc_scores = []
prec_scores = []
rec_scores = []
f1_scores = []

for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]

    pipeline.fit(X_train, y_train)
    y_pred = pipeline.predict(X_test)

    mae = mean_absolute_error(y_test, y_pred)
    mae_scores.append(mae)

    rmse = np.sqrt(mean_squared_error(y_test, y_pred))
    rmse_scores.append(rmse)
    
    r2 = r2_score(y_test, y_pred)
    r2_scores.append(r2)
    
    y_pred_disc = pd.cut(y_pred, bins=bins, labels=labels, include_lowest=True)
    y_test_disc = pd.cut(y_test, bins=bins, labels=labels, include_lowest=True)
    
    mask = (~pd.isna(y_pred_disc)) & (~pd.isna(y_test_disc))
    y_pred_disc = y_pred_disc[mask]
    y_test_disc = y_test_disc[mask]

    acc = accuracy_score(y_test_disc, y_pred_disc)
    prec = precision_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    rec = recall_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    f1 = f1_score(y_test_disc, y_pred_disc, average='macro', zero_division=0)
    
    acc_scores.append(acc)
    prec_scores.append(prec)
    rec_scores.append(rec)
    f1_scores.append(f1)


print("\nMean Absolute Error (per fold):", mae_scores)
print("Mean MAE:", np.mean(mae_scores))

print("\nRoot Mean Squared Error (per fold):", rmse_scores)
print("Mean RMSE:", np.mean(rmse_scores))

print("\nR2 - Coefficient of Determination (per fold):", r2_scores)
print("Mean R2:", np.mean(r2_scores))

print("\n--- Summary after discretization---")
print(f"Mean Accuracy: {np.mean(acc_scores):.4f}")
print(f"Mean Precision: {np.mean(prec_scores):.4f}")
print(f"Mean Recall: {np.mean(rec_scores):.4f}")
print(f"Mean F1-score: {np.mean(f1_scores):.4f}")

# saving resoults to df
df_results = pd.DataFrame({
    'MAE': mae_scores,
    'RMSE': rmse_scores,
    'R2': r2_scores
})
df_results.to_csv("3a_regression_no_resample_MLP.csv", index=False)

# saving results to df (for comparison with classification)
df_results = pd.DataFrame({
    'acc': acc_scores,
    'prec': prec_scores,
    'rec': rec_scores,
    'f1': f1_scores
})
df_results.to_csv("3a_classification_no_resample_MLP.csv", index=False)
